feat: AWS credentials lib for AWS signature
feat: read & write credentials for AWS Signature
chore: add todo
fix: remove utils.js from default.conf
fix: comment for readCredentials()
fix: aws credentials in default.conf
fix: aws_credentials.js test
fix: end of if statement
fix: unit test for s3gateway and aws credentials
fix: unit test for session token
feat: reusable function for unit test w/ w/o session token
fix: remove unnecessary comments

chore: add description for NJS files
